requirements.txt
boto==2.42.0
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pilkit==1.1.13
Pillow==4.2.1
py==1.4.31
PyYAML==3.11
requests==2.10.0
Werkzeug==0.12.1
